src/app/dashboard/transaction-review/transaction-review.component.ts
import { Component, Inject, OnInit } from '@angular/core';
import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
import { TransactionFormData } from '../../shared/models/transaction-form-data.model';

@Component({
  selector: 'app-transaction-review',
  templateUrl: './transaction-review.component.html',
  styleUrls: ['./transaction-review.component.scss']
})
export class TransactionReviewComponent implements OnInit {

  constructor(public dialogRef: MatDialogRef<TransactionReviewComponent>,
    @Inject(MAT_DIALOG_DATA) public data: TransactionFormData) {
  }

  ngOnInit(): void {
  }

  dismiss(): void {
    this.dialogRef.close(false);
  }

  confirm(): void {
    this.dialogRef.close(true);
  }

}
